+++ ports/emlib/wdog_software_watchdog.c
@@ -17,33 +17,34 @@
 //! - The WDOG peripheral supports discrete timeouts. This port chooses the closest period
 //!   that is less than or equal to the requested value. Valid selections for
 //!   MEMFAULT_WATCHDOG_SW_TIMEOUT_SECS range from 1s - 128s.
-
 #include "memfault/ports/watchdog.h"
 
+// non-module includes below
+#include "em_cmu.h"
+#include "em_wdog.h"
 #include "memfault/config.h"
 #include "memfault/core/debug_log.h"
 
-#include "em_wdog.h"
-#include "em_cmu.h"
-
 #if defined(WDOG0)
-#define MEMFAULT_WDOG WDOG0
-#define MEMFAULT_WDOG_IRQn WDOG0_IRQn
+  #define MEMFAULT_WDOG WDOG0
+  #define MEMFAULT_WDOG_IRQn WDOG0_IRQn
 
-#ifndef MEMFAULT_EXC_HANDLER_WATCHDOG
-#  error "Port expects following define to be set: -DMEMFAULT_EXC_HANDLER_WATCHDOG=WDOG0_IRQHandler"
-#endif
+  #ifndef MEMFAULT_EXC_HANDLER_WATCHDOG
+    #error \
+      "Port expects following define to be set: -DMEMFAULT_EXC_HANDLER_WATCHDOG=WDOG0_IRQHandler"
+  #endif
 
 #elif defined(WDOG)
-#define MEMFAULT_WDOG WDOG
-#define MEMFAULT_WDOG_IRQn WDOG_IRQn
+  #define MEMFAULT_WDOG WDOG
+  #define MEMFAULT_WDOG_IRQn WDOG_IRQn
 
-#ifndef MEMFAULT_EXC_HANDLER_WATCHDOG
-#  error "Port expects following define to be set: -DMEMFAULT_EXC_HANDLER_WATCHDOG=WDOG_IRQHandler"
-#endif
+  #ifndef MEMFAULT_EXC_HANDLER_WATCHDOG
+    #error \
+      "Port expects following define to be set: -DMEMFAULT_EXC_HANDLER_WATCHDOG=WDOG_IRQHandler"
+  #endif
 
 #else
-# error "Could not find an available WDOG to use"
+  #error "Could not find an available WDOG to use"
 #endif
 
 #define MEMFAULT_EM_WDOG_FREQ_HZ 1000
@@ -55,6 +56,7 @@
 
 #define MEMFAULT_EM_WDOG_MAX_TIMEOUT_MS (MEMFAULT_EM_WDOG_PERSEL_TO_TIMEOUT_MS(wdogPeriod_256k))
 
+// clang-format off
 static void prv_build_configuration(WDOG_Init_TypeDef *cfg, uint32_t persel) {
   *cfg = (WDOG_Init_TypeDef) {
     .enable = true,
@@ -70,8 +72,6 @@ static void prv_build_configuration(WDOG_Init_TypeDef *cfg, uint32_t persel) {
 #if defined(_WDOG_CTRL_CLKSEL_MASK)
     // use internal 1kHz clock for largest range
     .clkSel = wdogClkSelULFRCO,
-#else
-    #error "Port doesn't support WDOG Variant - cannot configure clock"
 #endif
     .perSel = persel,
 
@@ -91,23 +91,24 @@ static void prv_build_configuration(WDOG_Init_TypeDef *cfg, uint32_t persel) {
 #endif
   };
 }
+// clang-format on
 
 static int prv_configure_watchdog_with_timeout(uint32_t timeout_ms) {
-
   // NB: An interrupt can be configured to fire at 25%, 50%, or 75% of the watchdog cycle
   // configured. We'll use this interrupt as our "software watchdog" and configure it to be at
   // the 50% interval
-  const size_t sw_wdog_max_ms = MEMFAULT_EM_WDOG_WARNING_TIMEOUT_MS(MEMFAULT_EM_WDOG_MAX_TIMEOUT_MS);
+  const size_t sw_wdog_max_ms =
+    MEMFAULT_EM_WDOG_WARNING_TIMEOUT_MS(MEMFAULT_EM_WDOG_MAX_TIMEOUT_MS);
   if (timeout_ms > sw_wdog_max_ms) {
-    MEMFAULT_LOG_ERROR("Requested wdog timeout (%d) exceeds max supported (%d)",
-                       (int)timeout_ms, (int)sw_wdog_max_ms);
+    MEMFAULT_LOG_ERROR("Requested wdog timeout (%d) exceeds max supported (%d)", (int)timeout_ms,
+                       (int)sw_wdog_max_ms);
     return -1;
   }
 
   uint32_t persel = wdogPeriod_9;
   while (persel <= wdogPeriod_256k) {
     const uint32_t warning_timeout_ms =
-        MEMFAULT_EM_WDOG_WARNING_TIMEOUT_MS(MEMFAULT_EM_WDOG_PERSEL_TO_TIMEOUT_MS(persel));
+      MEMFAULT_EM_WDOG_WARNING_TIMEOUT_MS(MEMFAULT_EM_WDOG_PERSEL_TO_TIMEOUT_MS(persel));
     if (warning_timeout_ms > timeout_ms) {
       break;
     }
@@ -115,7 +116,7 @@ static int prv_configure_watchdog_with_timeout(uint32_t timeout_ms) {
   }
 
   // choose the closest timeout without going over the desired max
-  persel = (persel == 0)  ? 0 : persel - 1;
+  persel = (persel == 0) ? 0 : persel - 1;
 
   const size_t actual_timeout_ms = MEMFAULT_EM_WDOG_PERSEL_TO_TIMEOUT_MS(persel);
   MEMFAULT_LOG_DEBUG("Configuring SW Watchdog. SW Timeout=%dms HW Timeout=%dms",
@@ -123,9 +124,11 @@ static int prv_configure_watchdog_with_timeout(uint32_t timeout_ms) {
                      actual_timeout_ms);
 
   // Low energy clock must be on to use the watchdog
-  if ((CMU->HFBUSCLKEN0 & CMU_HFBUSCLKEN0_LE) == 0) {
-    CMU->HFBUSCLKEN0 |= CMU_HFBUSCLKEN0_LE;
-  }
+#if defined(_SILICON_LABS_32B_SERIES_2)
+  CMU_ClockEnable(cmuClock_WDOG0, true);
+#else   // !defined(_SILICON_LABS_32B_SERIES_2)
+  CMU_ClockEnable(cmuClock_HFLE, true);
+#endif  // defined(_SILICON_LABS_32B_SERIES_2)
 
   if (WDOGn_IsLocked(MEMFAULT_WDOG)) {
     MEMFAULT_LOG_ERROR("Watchdog is locked and cannot be reconfigured");
@@ -137,13 +140,18 @@ static int prv_configure_watchdog_with_timeout(uint32_t timeout_ms) {
   NVIC_DisableIRQ(MEMFAULT_WDOG_IRQn);
   NVIC_ClearPendingIRQ(MEMFAULT_WDOG_IRQn);
 
+  // Series 2 chips select the watchdog source via the CMU
+#if defined(_SILICON_LABS_32B_SERIES_2)
+  CMU_ClockSelectSet(cmuClock_WDOG0, cmuSelect_ULFRCO);
+#endif  // defined(_SILICON_LABS_32B_SERIES_2)
+
   WDOG_Init_TypeDef cfg;
   prv_build_configuration(&cfg, persel);
   WDOGn_Init(MEMFAULT_WDOG, &cfg);
 
   // enable the warning interrupt. This will be used to capture a coredump rather than just letting
   // the hardware watchdog immediately reboot the system
-  const uint32_t warn_int_mask = 0x2; // "WARN Interrupt Enable"
+  const uint32_t warn_int_mask = 0x2;  // "WARN Interrupt Enable"
   WDOGn_IntClear(MEMFAULT_WDOG, warn_int_mask);
   WDOGn_IntEnable(MEMFAULT_WDOG, warn_int_mask);
 
@@ -154,7 +162,6 @@ static int prv_configure_watchdog_with_timeout(uint32_t timeout_ms) {
 
   // finally, with everything setup, start the watchdog!
   WDOGn_Enable(MEMFAULT_WDOG, true);
-  WDOGn_SyncWait(MEMFAULT_WDOG);
   return 0;
 }
 
@@ -182,6 +189,6 @@ int memfault_software_watchdog_feed(void) {
   // Enteriung EM2 or EM3 power states while this is in progress will
   // cause the operation to be aborted. WDOGn_SyncWait() can be used to
   // block until the operation is complete!
-  WDOG_Feed();
+  WDOGn_Feed(MEMFAULT_WDOG);
   return 0;
 }