+++ scripts/mflt-build-id/tasks/__init__.py
@@ -3,15 +3,6 @@
 # See License.txt for details
 #
 
-# flake8: noqa: M900
+from mflt.testing.task_templates import make_test_task
 
-# We don't have invoke in the requirements because this is used by *other* packages,
-# not by this package itself.
-
-import invoke
-
-
-@invoke.task
-def test(ctx):
-    """Run tests."""
-    ctx.run("pytest", pty=True)
+test = make_test_task()