+++ tasks/nrf.py
@@ -39,7 +39,7 @@ def run_arm_toolchain_check(ctx):
             )
         )
 
-        raise Exception(msg)
+        raise FileNotFoundError(msg)
 
 
 @task